NF: DeckDueTreeNode in its own file
It seems to be complex enough and related to deck picker, so it makes sens to move it away
